Update dependencies and convert to TypeScript
