+++ js/package.json
@@ -1,13 +1,19 @@
 {
   "private": true,
   "name": "@flarum/akismet",
+  "prettier": "@flarum/prettier-config",
   "dependencies": {
     "flarum-webpack-config": "^1.0.0",
     "webpack": "^4.46.0",
-    "webpack-cli": "^4.9.0"
+    "webpack-cli": "^4.9.1"
   },
   "scripts": {
     "dev": "webpack --mode development --watch",
     "build": "webpack --mode production"
+  },
+  "devDependencies": {
+    "@flarum/prettier-config": "^1.0.0",
+    "flarum-tsconfig": "^1.0.2",
+    "prettier": "^2.5.1"
   }
 }
