+++ js/src/admin/index.ts
@@ -1,0 +1,9 @@
+import app from 'flarum/admin/app';
+
+app.initializers.add('flarum-akismet', () => {
+  app.extensionData.for('flarum-akismet').registerSetting({
+    setting: 'flarum-akismet.api_key',
+    type: 'text',
+    label: app.translator.trans('flarum-akismet.admin.akismet_settings.api_key_label'),
+  });
+});
